+++ src/compas_model/materials/__init__.py
@@ -1,0 +1,4 @@
+from .material import Material
+from .material import ElasticIsotropic
+
+__all__ = ["Material", "ElasticIsotropic"]