+++ src/compas_model/materials/material.py
@@ -6,31 +6,28 @@ class Material(Data):
 
     Parameters
     ----------
-    density : float
+    name : str
+        Name of the material.
+    rho : float
         Density of the material.
-    expansion : float, optional
+    eps : float, optional
         Thermal expansion coefficient, by default None.
 
     Attributes
     ----------
-    density : float
+    name : str
+        Name of the material.
+    rho : float
         Density of the material.
-    expansion : float
+    eps : float
         Thermal expansion coefficient.
-    key : int
-        The key index of the material. It is automatically assigned to material
-        once it is added to the model.
     """
 
-    def __init__(self, *, density, expansion=None, **kwargs):
+    def __init__(self, *, name, rho, eps=None, **kwargs):
         super(Material, self).__init__(**kwargs)
-        self.density = density
-        self.expansion = expansion
-        self._key = None
-
-    @property
-    def key(self):
-        return self._key
+        self.name = name
+        self.rho = rho
+        self.eps = eps
 
     def __str__(self):
         return """
@@ -40,36 +37,47 @@ name        : {}
 density     : {}
 expansion   : {}
 """.format(
-            self.__class__.__name__, len(self.__class__.__name__) * "-", self.name, self.density, self.expansion
+            self.__class__.__name__, len(self.__class__.__name__) * "-", self.name, self.rho, self.eps
         )
 
 
 # ==============================================================================
 # linear elastic
 # ==============================================================================
+
+
 class ElasticIsotropic(Material):
-    """Elastic, isotropic and homogeneous material
+    """Elastic, isotropic and homogeneous material.
 
     Parameters
     ----------
+    name : str
+        Name of the material.
     E : float
         Young's modulus E.
     v : float
         Poisson's ratio v.
+    rho : float
+        Density of the material.
+    eps : float, optional
+        Thermal expansion coefficient, by default None.
 
     Attributes
     ----------
+    name : str
+        Name of the material.
     E : float
         Young's modulus E.
     v : float
         Poisson's ratio v.
-    G : float
-        Shear modulus (automatically computed from E and v)
+    G : float, read-only
+        Shear modulus (automatically computed from E and v).
 
     """
 
-    def __init__(self, *, E, v, density, expansion=None, name=None, **kwargs):
-        super(ElasticIsotropic, self).__init__(density=density, expansion=expansion, name=name, **kwargs)
+    def __init__(self, *, name, E, v, rho, eps=None, **kwargs):
+        super(ElasticIsotropic, self).__init__(name=name, rho=rho, eps=eps, **kwargs)
+        self.name = name
         self.E = E
         self.v = v
 
@@ -85,7 +93,7 @@ E : {}
 v : {}
 G : {}
 """.format(
-            self.name, self.density, self.expansion, self.E, self.v, self.G
+            self.name, self.rho, self.eps, self.E, self.v, self.G
         )
 
     @property